requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
Flask==2.1.2
Flask-BasicAuth==0.2.0
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
MarkupSafe==2.1.1
MLB-StatsAPI==1.4.2
numpy==1.23.0
pandas==1.4.3
psycopg2-binary==2.9.3
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.1
scikit-learn==1.1.1
scipy==1.8.1
six==1.16.0
SQLAlchemy==1.4.39
threadpoolctl==3.1.0
urllib3==1.26.9
Werkzeug==2.1.2
zipp==3.8.0
